nassh: add a workaround for missing chrome.runtime bug
Sometimes when the window loads, the chrome.runtime object (among
others) isn't populated.  This should never happen as it's part of
Chrome's extension API to have it always available.  Refreshing the
page seems to workaround whatever the problem is and all the APIs
are usable again.

Unfortunately, these missing APIs break Secure Shell and users only
see a black window.  So add a hack to recover until we can figure
out better what's going on.

Bug: chromium:924656
Change-Id: Ie2cd61a44c9771f0c775841fe51c7129b3367e86
Reviewed-on: https://chromium-review.googlesource.com/c/1434575
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>