nassh: fsp: rewrite write file loop
With the increase in write packet sizes, we can exceed the JS stack
limit and fail to actually copy out files.  This is because of the
use of String.fromCharCode.apply which is implemented as a recursive
call stack, and we pass it an array that can have 200k+ entries.

First, lets replace that with a call to lib.codec.codeUnitToString
which is not recursive, so it'll work regardless of the input sizes.
It should also have more constant performance (for smaller arrays,
it might be slower, but we tend to pass in large arrays more often).

Next, instead of using ArrayBuffer's .slice method to chunk up the
data (which creates a copy of the data), use typed arrays and their
.subarray method so we only ever have one copy of the buffer.

Bug: chromium:924764
Change-Id: I274ea8add0158798b88d182c05b4447f1c92e48f
Reviewed-on: https://chromium-review.googlesource.com/c/1437526
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>