nassh: sftp: split sftp stream logic out of tty
Using the Stream.Tty class for sftp clients has always been a hack:
we assume that it only ever calls writeUTF8, so we defined that in
the nassh.sftp.Client class and switched the Terminal.IO object to
that.  This makes the code more complicated & fragile than it needs
to be, so create a proper Stream object for sftp clients.

Change-Id: Iab67f13b40d80b9a9dc68f57660bf2a3433fc7c7
Reviewed-on: https://chromium-review.googlesource.com/c/1453977
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>