nassh: sftp: change DataPackets to data APIs
The string type implies encoding on the data (which we're assuming is
UTF-8) while the data type does not.  Change the DataPacket API to use
getData instead of getString.  This doesn't make a difference here, but
will matter after follow up commits.

We do the same in writeChunk that is creating a DataPacket on the fly.

Change-Id: I6e572a7e4643f0fcb89abe729a3d29fe674bf897
Reviewed-on: https://chromium-review.googlesource.com/c/1461364
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>