nassh: Add support for ECDSA keys via OpenPGP applet
Starting with version 3.0, the OpenPGP applet on smart cards support ECC
keys. With this commit, support for SSH authentication using any of the
three required NIST curves (nistp256, nistp384 and nistp521) on an
OpenPGP-enabled smart card is added to the 'gsc' agent backend.

The commit is structured in a way that allows support for other ECDSA
curves, such as the optional NIST curves, to be added simply by
appending their OIDs, identifiers, public key prefixes and hash
algorithms to the object nassh.agent.messages.OidToCurveInfo.

Change-Id: Ibb1b160a3341164911aa05c3263b9cefa4d62c4a