nassh: Bundle licenses of prod NPM dependencies
Provides a python script that collects the license file contents of all
NPM dependencies and exports them as a libdot resource.

The libapps NPM dependencies are divided into dev and prod dependencies
depending on whether their code will be part of the extension in the
Chrome Web Store or not.

Change-Id: I592fd0f03cde535b9c8d53ad8756db2d6d9c1d43